rc/meson-ir: we don't need to capture pulse duration. If we are decoding, then the edge is a pulse. Otherwise, it's a space.
The upstream meson-ir driver will determine the pulse duration from reading the decoder input register and multiplying it by the tickrate. But we found that this isn't very good. This is likely because AML devices tend to use NEC remotes. Even then, I'm not sure how reliable that is. They usually use a binary called 'amremote' which handles the remote decoding. I've never really looked in to it as it doesn't apply to us.

There's also another register which can be used to determine whether the hardware is decoding a pulse or not. This is better.

We use ir_raw_event_store_edge() instead of ir_raw_event_store(). This is more appropriate because there isn't an actual register for
providing duration directly. Instead we only get an interrupt on state change. So instead we should inform the kernel of the beginning / end of an IR pulse or
a space and let it do the rest.

Signed-off-by: Sam Nazarko <email@samnazarko.co.uk>