PD#151264: pinctrl: txhd: add support pinctrl & gpio for txhd
The patch add an new driver "pinctrl-meson.c" which is used for the chip
with the following update in the pinmux registers.

TXHD and later: use continuous 4-bit to choose function for per pin.

The new driver is ported based on the linux-4.9，the following features
have been inherited，and incompatible with the previous pinctrl driver.

1) pinmux configuration
2) pinconf configuration
3) gpio irq using
4) gpio ID definition

Change-Id: I838cbd1f7325dd3e1a61d5de8c8c6d6d2ea949cd
Signed-off-by: xingyu.chen <xingyu.chen@amlogic.com>
Signed-off-by: Bo Yang <bo.yang@amlogic.com>