OSMT #387075: Fix Ethernet stability in quiet environments.
This issue typically occurs in a network environment that does not feature a DHCP server or Avahi users.
The original AML PD was introduced to resolve issues with GXL boards that had Ethernet issues.
This was actually caused by the power core problem on a specific batch of chips.
Original issue does not affect our devices; but if there are a number of TX packets before an RX broadcast on any device, this fix will cause adverse problems.

Signed-off-by: Sam Nazarko <email@samnazarko.co.uk>